Add missing NVDA minor versions (#24)
Found while reviewing nvaccess/nvda#15888

Issue
NVDA API version json file does not cover all existing NVDA versions. These ones minor versions are missing:

2018.4.1
2019.1.1
2019.2.1
2019.3.1
I doubt any add-on is impacted by these missing minor versions. But for completeness and clarity, it's better to have them covered.

Solution
Fix the description of API 0.0.0 so that 2018.4.1 is covered
Added API for 2019.1.1 and 2019.2.1
Fixed description of 2019.3 to include also 2019.3.1. No API is added for this version since it contains only new translations; it is worth noting that this version is not listed in NVDA's change log.